Add single api request system to gradually replace current system
The api system was restricting the ability to add new api requests due to the various needs for each api request type being incompatible with each other. This was fixed by abstracting the common segments of the api requests into a single class to be used whenever api requests are needed.