Modify api system to be more versatile, configurable, and decoupled from components
Many components required access to the backend api and would make use of a separate api component to do so. This meant that each new component that needed to access the backend would require a separate api processing component. This pattern would inevitably lead to code that was duplicated, higher maintenance, and generally harder to debug.

To fix this, the api processing system was modified to be accessible in a single way to all components. Notably, the focus during this refactoring was to standardize the use of the api system while maintaining the unique needs of each api consumer. This was done by first abstracting the common api actions into newly created components. Then combining the entry and exit points of api processes into a singular (ApiResponse) component and making that a child of the api consumer. The standardized api processing is done outside the boundaries of the consumers, but the consumers can modify the default configuration of each api request to ensure there specific conditions and tasks are met.