Modify refresh and jwt usage
Due to a lack of persistence of the jwt after page reload, the actual jwt is now sent as a HttpOnly cookie. The jwt uses in the frontend have therefore been removed. To facilitate a more efficient refresh of the jwt when needed, a direct refresh request has been added to the api request component.

Furthermore, the logged in state boolean will be stored in local storage as a convenient way to alert the frontend to refresh the jwt if logged in.