Add sidebar and credential manager
Adding exchange credentials was coupled with adding a card which leads to less intuitive user experience. This was fixed by separating exchange addition into a sidebar that could be popped out on the home screen whenever the user would like to add an exchange.

Modify checkbox widget to use sock js and websockets

The checkbox widget fetches user data from the backend as changes occur to the user account. It was logical then to create websocket access for the checkbox widget via sock js to provide instant updates. Furthermore, the sockjs implementation has been modified to be centered in one location (within the app.js file) and passed to components via react context pattern.